refactor: 레이어별 클래스 분리
