+++ README.md
@@ -37,6 +37,9 @@
 - [x] 예약 추가 쿼리를 수정한다.
 - [x] 예약 조회 쿼리를 수정한다.
 
+- [x] 레이어드 아키텍처를 적용하여 레이어별 책임과 역할에 따라 클래스 분리한다.
+- [x] 분리한 클래스는 매번 새로 생성하지 않고 스프링 빈으로 등록해서 사용한다.
+
 ## 커밋 컨밴션
 - feat, docs, fix, refactor, test, style, chore
 > EX) feat: Piece id 필드 추가
